Refactor SX1301 / SX1302 GNSS handling.
This re-implements the GNSS handling directly into the libconcentratord
crate. This fixes an issue where the GNSS timestamps and concentrator
counter could run out of sync, causing TOO_EARLY / TOO_LATE errors on
GNSS based downlink scheduling.

The re-implementation also makes it easier to add partial GNSS support
for non-Ublox GNSS modules in case a precise time since GPS epoch is not
needed (e.g. when we are only interested in the location and an UTC
timestamp from the NMEA records).

This re-implementation does at the moment not calculate the XTAL
correction. In the old implementation, the Concentratord would calculate
it, but the calculated value was not used.

+++ chirpstack-concentratord-sx1302/src/config/mod.rs
@@ -77,7 +77,7 @@ impl Default for Api {
 pub struct Location {
     pub latitude: f64,
     pub longitude: f64,
-    pub altitude: i16,
+    pub altitude: f32,
 }
 
 #[derive(Serialize, Deserialize)]
